# docs/assets/openapi-spec.yml
openapi: 3.0.0
servers:
  - description: Agent Backchannel
    url: http://localhost:{port}
    variables:
      port:
        description: >
          The port to use for the backchannel
            * Acme:    9020
            * Bob:      9030
            * Faber:    9040
            * Mallory:  9050
        enum:
          - "9020"
          - "9030"
          - "9040"
          - "9050"
        default: "9020"
info:
  description: |
    This page documents the backchannel API the test harness uses to communicate with agents under tests.

    For more information checkout the [Aries Interoperability Information](http://aries-interop.info) page.
  version: "1.0.0"
  title: Aries Agent Test Harness Backchannel API
  contact:
    name: Province of British Columbia
    url: https://github.com/hyperledger/aries-agent-test-harness
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  x-logo:
    url: https://raw.githubusercontent.com/hyperledger/aries-agent-test-harness/master/docs/assets/aries-logo.png
    altText: Hyperledger Aries logo
...
paths:
...
  /agent/command/issue-credential/{credentialExchangeThreadId}:
    get:
      tags:
        - Issue Credential
      summary: Get credential exchange record by thread id
      operationId: IssueCredentialGetByThreadId
      parameters:
        - in: path
          name: credentialExchangeThreadId
          required: true
          schema:
            $ref: "#/components/schemas/ThreadId"
      responses:
        200:
          description: Credential Exchange
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCredentialOperationResponse"
        404:
          description: Credential exchange record not found
  /agent/command/issue-credential/send-proposal:
    post:
      tags:
        - Issue Credential
      summary: Send credential proposal
      description: >
        Send a `propose-credential` message to connection with `connection_id` in body.
      externalDocs:
        url: https://github.com/hyperledger/aries-rfcs/tree/master/features/0036-issue-credential#propose-credential
      operationId: IssueCredentialSendProposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: object
                  required:
                    - connection_id
                  properties:
                    comment:
                      $ref: "#/components/schemas/Comment"
                    credential_proposal:
                      $ref: "#/components/schemas/CredentialPreview"
                    schema_issuer_did:
                      $ref: "#/components/schemas/Did"
                    schema_id:
                      $ref: "#/components/schemas/SchemaId"
                    schema_name:
                      $ref: "#/components/schemas/SchemaName"
                    schema_version:
                      $ref: "#/components/schemas/SchemaVersion"
                    cred_def_id:
                      $ref: "#/components/schemas/CredentialDefinitionId"
                    issuer_did:
                      $ref: "#/components/schemas/Did"
                    connection_id:
                      $ref: "#/components/schemas/ConnectionId"
      responses:
        200:
          description: Credential proposal sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IssueCredentialOperationResponse"
                  - properties:
                      state:
                        example: proposal-sent
  /agent/command/issue-credential/send-offer:
    post:
      tags:
        - Issue Credential
      summary: Send credential offer
      description: >
        Send an `offer-credential` message.
      externalDocs:
        url: https://github.com/hyperledger/aries-rfcs/tree/master/features/0036-issue-credential#offer-credential
      operationId: IssueCredentialSendOffer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf:
                - title: Send Offer to start credential issuance flow
                  required:
                    - data
                  properties:
                    data:
                      type: object
                      required:
                        - connection_id
                        - credential_preview
                      properties:
                        comment:
                          $ref: "#/components/schemas/Comment"
                        credential_preview:
                          $ref: "#/components/schemas/CredentialPreview"
                        cred_def_id:
                          $ref: "#/components/schemas/CredentialDefinitionId"
                        connection_id:
                          $ref: "#/components/schemas/ConnectionId"
                - title: Send Offer in response to Proposal
                  required:
                    - id
                  properties:
                    id:
                      $ref: "#/components/schemas/ThreadId"
            examples:
              new:
                summary: Send Offer to start credential issuance flow
                description: >
                  When sending an offer to start the credential issuance flow, the data
                  of the credential that is going to be issued must be passed as value to the data key.
                value:
                  data:
                    comment:
                      $ref: "#/components/schemas/Comment/example"
                    credential_preview:
                      $ref: "#/components/schemas/CredentialPreview/example"
                    cred_def_id:
                      $ref: "#/components/schemas/CredentialDefinitionId/example"
                    connection_id:
                      $ref: "#/components/schemas/ConnectionId/example"
              response:
                summary: Send Offer in response to Proposal
                description: >
                  When sending an offer in response to a received proposal only the
                  thread id should be passed
                value:
                  id:
                    $ref: "#/components/schemas/ThreadId/example"
      responses:
        200:
          description: Credential offer sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IssueCredentialOperationResponse"
                  - properties:
                      state:
                        example: offer-sent
  /agent/command/issue-credential/send-request:
    post:
      tags:
        - Issue Credential
      summary: Send credential request
      externalDocs:
        url: https://github.com/hyperledger/aries-rfcs/tree/master/features/0036-issue-credential#request-credential
      operationId: IssueCredentialSendRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              anyOf:
                - title: Send Request to start credential issuance flow
                  required:
                    - id
                  properties:
                    id:
                      $ref: "#/components/schemas/ConnectionId"
                - title: Send Request in response to Offer
                  required:
                    - id
                  properties:
                    id:
                      $ref: "#/components/schemas/ThreadId"
            examples:
              response:
                summary: Send Request in response to Offer
                description: >
                  When sending a Request in response to a received offer only the
                  thread id should be passed
                value:
                  id:
                    $ref: "#/components/schemas/ThreadId/example"
              new:
                summary: Send Request to start credential issuance flow
                description: >
                  This is not possible with Indy based credentials.
                value:
                  id:
                    $ref: "#/components/schemas/ConnectionId/example"
      responses:
        200:
          description: Credential request sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IssueCredentialOperationResponse"
                  - properties:
                      state:
                        example: request-sent
  /agent/command/issue-credential/issue:
    post:
      tags:
        - Issue Credential
      summary: Issue Credential
      externalDocs:
        url: https://github.com/hyperledger/aries-rfcs/tree/master/features/0036-issue-credential#issue-credential
      operationId: IssueCredentialIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ThreadId"
                data:
                  type: object
                  properties:
                    comment:
                      $ref: "#/components/schemas/Comment"
      responses:
        200:
          description: Credential issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IssueCredentialOperationResponse"
                  - properties:
                      state:
                        example: credential-issued
  /agent/command/issue-credential/store:
    post:
      tags:
        - Issue Credential
      summary: Store Credential
      operationId: IssueCredentialStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ThreadId"
      responses:
        200:
          description: Credential stored
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IssueCredentialOperationResponse"
                  - required:
                      - credential_id
                    properties:
                      state:
                        example: done
...
  /agent/command/credential/{credentialId}:
    get:
      tags:
        - Credential
      summary: Get credential by id
      operationId: CredentialGetById
      parameters:
        - in: path
          name: credentialId
          required: true
          schema:
            $ref: "#/components/schemas/CredentialId"
      responses:
        200:
          description: Credential
          content:
            application/json:
              schema:
                oneOf:
                  - title: Indy Credential
                    description: Expected response for an Indy credential
                    required:
                      - referent
                      - schema_id
                      - cred_def_id
                    properties:
                      referent:
                        $ref: "#/components/schemas/CredentialId"
                      schema_id:
                        $ref: "#/components/schemas/SchemaId"
                      cred_def_id:
                        $ref: "#/components/schemas/CredentialDefinitionId"
                  - title: W3C Credential
                    description: Expected response for a W3C credential
                    required:
                      - credential_id
                      - credential
                    properties:
                      credential_id:
                        $ref: "#/components/schemas/CredentialId"
                      credential:
                        $ref: "#/components/schemas/W3CCredential"
        404:
          description: Credential not found
...
components:
  schemas:
...
    ConnectionId:
      title: Connection Id
      example: "8fb9ea21-d094-4506-86b6-c7c1627d753a"
      type: string
    ThreadId:
      title: Thread Id
      example: "e7280776-5315-4569-8cac-42cae108adfe"
      type: string
    SchemaId:
      title: Schema Id
      example: "Y9oNbFNTgxrRuvxQk3xEzr:2:test_schema:1.0.0"
      type: string
    CredentialDefinitionId:
      title: Credential Definition Id
      example: "TTK1F3HKKBdtb6HrhrxYtC:3:CL:12:default"
      type: string
...
    Did:
      title: DID
      example: Y9oNbFNTgxrRuvxQk3xEzr
      type: string
...
    Comment:
      title: Comment
      example: Hello World
      type: string
    SchemaName:
      title: Schema Name
      type: string
      example: test_schema
    SchemaVersion:
      title: Schema Version
      type: string
      example: "1.0.0"
...
    W3CCredential:
      title: W3C Verifiable Credential
      type: object
      externalDocs:
        url: https://www.w3.org/TR/vc-data-model/
      example:
        {
          "@context":
            [
              "https://www.w3.org/2018/credentials/v1",
              "https://www.w3.org/2018/credentials/examples/v1",
            ],
          "id": "http://example.edu/credentials/1872",
          "type": ["VerifiableCredential", "AlumniCredential"],
          "issuer": "https://example.edu/issuers/565049",
          "issuanceDate": "2010-01-01T19:73:24Z",
          "credentialSubject":
            {
              "id": "did:example:ebfeb1f712ebc6f1c276e12ec21",
              "alumniOf":
                {
                  "id": "did:example:c276e12ec21ebfeb1f712ebc6f1",
                  "name":
                    [
                      { "value": "Example University", "lang": "en" },
                      { "value": "Exemple d'Université", "lang": "fr" },
                    ],
                },
            },
          "proof":
            {
              "type": "RsaSignature2018",
              "created": "2017-06-18T21:19:10Z",
              "proofPurpose": "assertionMethod",
              "verificationMethod": "https://example.edu/issuers/keys/1",
              "jws":
                "eyJhbGciOiJSUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..TCYt5X
                sITJX1CxPCT8yAV-TVkIEq_PbChOMqsLfRoPsnsgw5WEuts01mq-pQy7UJiN5mgRxD-WUc
                X16dUEMGlv50aqzpqh4Qktb3rk-BuQy72IFLOqV0G_zS245-kronKb78cPN25DGlcTwLtj
                PAYuNzVBAh4vGHSrQyHUdBBPM",
            },
        }
    CredentialPreview:
      title: Credential Preview
      type: object
      required:
        - "@type"
        - attributes
      properties:
        "@type":
          type: string
          enum:
            - "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/issue-credential/1.0/credential-preview"
          example: "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/issue-credential/1.0/credential-preview"
        attributes:
          type: array
          items:
            type: object
            required:
              - name
              - value
            properties:
              name:
                type: string
              mime-type:
                type: string
              value:
                type: string
      example:
        {
          "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/issue-credential/1.0/credential-preview",
          "attributes":
            [
              { "name": "attr_1", "value": "value_1" },
              { "name": "attr_2", "value": "value_2" },
              { "name": "attr_3", "value": "10" },
            ],
        }
    IssueCredentialState:
      title: Issue Credential State
      description: All possible issue credential state values
      type: string
      example: proposal-sent
      enum:
        - proposal-sent
        - proposal-received
        - offer-sent
        - offer-received
        - request-sent
        - request-received
        - credential-issued
        - credential-received
        - done
...
    IssueCredentialOperationResponse:
      title: Issue Credential Operation Response
      type: object
      required:
        - state
        - thread_id
      properties:
        state:
          $ref: "#/components/schemas/IssueCredentialState"
        thread_id:
          $ref: "#/components/schemas/ThreadId"
        credential_id:
          $ref: "#/components/schemas/CredentialId"
    CredentialId:
      title: Credential Id
      example: "828761d2-65e9-44a5-852e-1f60b943ab65"
      type: string
...